+++ lib/src/reverse_engineering/responses/search_page.dart
@@ -3,12 +3,14 @@ import 'dart:convert';
 import 'package:collection/collection.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart' as parser;
+import 'package:youtube_explode_dart/src/search/search_channel.dart';
 
 import '../../../youtube_explode_dart.dart';
 import '../../extensions/helpers_extension.dart';
 import '../../retry.dart';
 import '../../search/base_search_content.dart';
 import '../../search/related_query.dart';
+import '../../search/search_filter.dart';
 import '../../search/search_video.dart';
 import '../../videos/videos.dart';
 import '../youtube_http_client.dart';
@@ -28,10 +30,7 @@ class SearchPage {
       return _initialData!;
     }
 
-    final scriptText = root!
-        .querySelectorAll('script')
-        .map((e) => e.text)
-        .toList(growable: false);
+    final scriptText = root!.querySelectorAll('script').map((e) => e.text).toList(growable: false);
     return scriptText.extractGenericData(
         (obj) => _InitialData(obj),
         () => TransientFailureException(
@@ -39,47 +38,36 @@ class SearchPage {
   }
 
   ///
-  SearchPage(this.root, this.queryString, [_InitialData? initialData])
-      : _initialData = initialData;
+  SearchPage(this.root, this.queryString, [_InitialData? initialData]) : _initialData = initialData;
 
   Future<SearchPage?> nextPage(YoutubeHttpClient httpClient) async {
-    if (initialData.continuationToken == '' ||
-        initialData.estimatedResults == 0) {
+    if (initialData.continuationToken == '' || initialData.estimatedResults == 0) {
       return null;
     }
     return get(httpClient, queryString, token: initialData.continuationToken);
   }
 
   ///
-  static Future<SearchPage> get(
-      YoutubeHttpClient httpClient, String queryString,
-      {String? token}) {
+  static Future<SearchPage> get(YoutubeHttpClient httpClient, String queryString,
+      {String? token, SearchFilter filter = const SearchFilter('')}) {
     if (token != null) {
-      var url =
-          'https://www.youtube.com/youtubei/v1/search?key=AIzaSyAO_FJ2SlqU8Q4STEHLGCilw_Y9_11qcW8';
+      var url = 'https://www.youtube.com/youtubei/v1/search?key=AIzaSyAO_FJ2SlqU8Q4STEHLGCilw_Y9_11qcW8';
 
       return retry(() async {
         var body = {
           'context': const {
-            'client': {
-              'hl': 'en',
-              'clientName': 'WEB',
-              'clientVersion': '2.20200911.04.00'
-            }
+            'client': {'hl': 'en', 'clientName': 'WEB', 'clientVersion': '2.20200911.04.00'}
           },
           'continuation': token
         };
 
-        var raw =
-            await httpClient.post(Uri.parse(url), body: json.encode(body));
-        return SearchPage(
-            null, queryString, _InitialData(json.decode(raw.body)));
+        var raw = await httpClient.post(Uri.parse(url), body: json.encode(body));
+        return SearchPage(null, queryString, _InitialData(json.decode(raw.body)));
       });
       // Ask for next page,
 
     }
-    var url =
-        'https://www.youtube.com/results?search_query=${Uri.encodeQueryComponent(queryString)}';
+    var url = 'https://www.youtube.com/results?search_query=${Uri.encodeQueryComponent(queryString)}&sp=${filter.value}';
     return retry(() async {
       var raw = await httpClient.getString(url);
       return SearchPage.parse(raw, queryString);
@@ -157,9 +145,7 @@ class _InitialData {
   }
 
   // Contains only [SearchVideo] or [SearchPlaylist]
-  late final List<BaseSearchContent> searchContent =
-      getContentContext()?.map(_parseContent).whereNotNull().toList() ??
-          const [];
+  late final List<BaseSearchContent> searchContent = getContentContext()?.map(_parseContent).whereNotNull().toList() ?? const [];
 
   List<RelatedQuery> get relatedQueries =>
       getContentContext()
@@ -167,10 +153,8 @@ class _InitialData {
           .map((e) => e.get('horizontalCardListRenderer')?.getList('cards'))
           .firstOrNull
           ?.map((e) => e['searchRefinementCardRenderer'])
-          .map((e) => RelatedQuery(
-              e.searchEndpoint.searchEndpoint.query,
-              VideoId(
-                  Uri.parse(e.thumbnail.thumbnails.first.url).pathSegments[1])))
+          .map((e) =>
+              RelatedQuery(e.searchEndpoint.searchEndpoint.query, VideoId(Uri.parse(e.thumbnail.thumbnails.first.url).pathSegments[1])))
           .toList()
           .cast<RelatedQuery>() ??
       const [];
@@ -178,11 +162,7 @@ class _InitialData {
   List<dynamic> get relatedVideos =>
       getContentContext()
           ?.where((e) => e['shelfRenderer'] != null)
-          .map((e) => e
-              .get('shelfRenderer')
-              ?.get('content')
-              ?.get('verticalListRenderer')
-              ?.getList('items'))
+          .map((e) => e.get('shelfRenderer')?.get('content')?.get('verticalListRenderer')?.getList('items'))
           .firstOrNull
           ?.map(_parseContent)
           .whereNotNull()
@@ -191,8 +171,7 @@ class _InitialData {
 
   late final String? continuationToken = _getContinuationToken();
 
-  late final int estimatedResults =
-      int.parse(root.getT<String>('estimatedResults') ?? '0');
+  late final int estimatedResults = int.parse(root.getT<String>('estimatedResults') ?? '0');
 
   BaseSearchContent? _parseContent(Map<String, dynamic>? content) {
     if (content == null) {
@@ -207,47 +186,32 @@ class _InitialData {
           _parseRuns(renderer.get('ownerText')?.getList('runs')),
           _parseRuns(renderer.get('descriptionSnippet')?.getList('runs')),
           renderer.get('lengthText')?.getT<String>('simpleText') ?? '',
-          int.parse(renderer
-                  .get('viewCountText')
-                  ?.getT<String>('simpleText')
-                  ?.stripNonDigits()
-                  .nullIfWhitespace ??
-              renderer
-                  .get('viewCountText')
-                  ?.getList('runs')
-                  ?.firstOrNull
-                  ?.getT<String>('text')
-                  ?.stripNonDigits()
-                  .nullIfWhitespace ??
+          int.parse(renderer.get('viewCountText')?.getT<String>('simpleText')?.stripNonDigits().nullIfWhitespace ??
+              renderer.get('viewCountText')?.getList('runs')?.firstOrNull?.getT<String>('text')?.stripNonDigits().nullIfWhitespace ??
               '0'),
           (renderer.get('thumbnail')?.getList('thumbnails') ?? const [])
-              .map((e) =>
-                  Thumbnail(Uri.parse(e['url']), e['height'], e['width']))
+              .map((e) => Thumbnail(Uri.parse(e['url']), e['height'], e['width']))
               .toList(),
           renderer.get('publishedTimeText')?.getT<String>('simpleText'),
-          renderer
-                  .get('viewCountText')
-                  ?.getList('runs')
-                  ?.elementAtSafe(1)
-                  ?.getT<String>('text')
-                  ?.trim() ==
-              'watching');
+          renderer.get('viewCountText')?.getList('runs')?.elementAtSafe(1)?.getT<String>('text')?.trim() == 'watching');
     }
     if (content['radioRenderer'] != null) {
       var renderer = content.get('radioRenderer')!;
 
-      return SearchPlaylist(
-          PlaylistId(renderer.getT<String>('playlistId')!),
+      return SearchPlaylist(PlaylistId(renderer.getT<String>('playlistId')!), renderer.get('title')!.getT<String>('simpleText')!,
+          int.parse(_parseRuns(renderer.get('videoCountText')?.getList('runs')).stripNonDigits().nullIfWhitespace ?? '0'));
+    }
+    if (content['channelRenderer'] != null) {
+      var renderer = content.get('channelRenderer')!;
+      return SearchChannel(
+          ChannelId(renderer.getT<String>('channelId')!),
           renderer.get('title')!.getT<String>('simpleText')!,
-          int.parse(_parseRuns(renderer.get('videoCountText')?.getList('runs'))
-                  .stripNonDigits()
-                  .nullIfWhitespace ??
-              '0'));
+          renderer.get('descriptionSnippet')?.getList('runs')?.parseRuns() ?? '',
+          renderer.get('videoCountText')!.getList('runs')!.first.getT<String>('text')!.parseInt()!);
     }
     // Here ignore 'horizontalCardListRenderer' & 'shelfRenderer'
     return null;
   }
 
-  String _parseRuns(List<dynamic>? runs) =>
-      runs?.map((e) => e['text']).join() ?? '';
+  String _parseRuns(List<dynamic>? runs) => runs?.map((e) => e['text']).join() ?? '';
 }