+++ lib/src/search/search_channel.dart
@@ -1,16 +1,29 @@
+import 'package:equatable/equatable.dart';
+
 import '../channels/channel_id.dart';
+import 'base_search_content.dart';
 
 /// Metadata related to a search query result (channel)
-class SearchChannel {
-  /// ChannelId.
-  final ChannelId channelId;
+class SearchChannel extends BaseSearchContent with EquatableMixin {
+  /// Channel id.
+  final ChannelId id;
 
   /// Channel name.
-  final String channelName;
+  final String name;
+
+  /// Description snippet.
+  /// Can be empty.
+  final String description;
+
+  /// Channel uploaded videos.
+  final int videoCount;
 
   /// Initialize a [SearchChannel] instance.
-  SearchChannel(this.channelId, this.channelName);
+  SearchChannel(this.id, this.name, this.description, this.videoCount);
 
   @override
-  String toString() => '(Channel) $channelName ($channelId)';
+  String toString() => '(Channel) $name ($id)';
+
+  @override
+  List<Object?> get props => [id, name, description, videoCount];
 }