+++ lib/src/search/search_client.dart
@@ -1,10 +1,12 @@
 import 'dart:convert';
 
 import '../../youtube_explode_dart.dart';
+import '../extensions/helpers_extension.dart';
 import '../retry.dart';
 import '../reverse_engineering/responses/search_page.dart';
 import '../reverse_engineering/youtube_http_client.dart';
 import 'base_search_content.dart';
+import 'search_filter.dart';
 import 'search_list.dart';
 import 'search_query.dart';
 
@@ -19,37 +21,40 @@ class SearchClient {
   /// (from the video search page).
   /// The videos are sent in batch of 20 videos.
   /// You [SearchList.nextPage] to get the next batch of videos.
-  Future<SearchList> getVideos(String searchQuery) {
-    var stream = getVideosFromPage(searchQuery).cast<SearchVideo>();
-    return SearchList.create(stream);
+  Future<SearchList> getVideos(String searchQuery, {SearchFilter filter = const SearchFilter('')}) async {
+    final page = await SearchPage.get(_httpClient, searchQuery, filter: filter);
+
+    return SearchList(
+        page.initialData.searchContent
+            .whereType<SearchVideo>()
+            .map((e) => Video(e.id, e.title, e.author, null, e.uploadDate?.toDateTime(), e.description, e.duration.toDuration(),
+                ThumbnailSet(e.id.value), null, Engagement(e.viewCount, null, null), e.isLive))
+            .toList(),
+        page,
+        _httpClient);
   }
 
   /// Enumerates videos returned by the specified search query
   /// (from the video search page).
   /// Contains only instances of [SearchVideo] or [SearchPlaylist]
   Stream<BaseSearchContent> getVideosFromPage(String searchQuery,
-      {bool onlyVideos = true}) async* {
-    SearchPage? page =
-        await retry(() async => SearchPage.get(_httpClient, searchQuery));
-    if (onlyVideos) {
-      yield* Stream.fromIterable(
-          page!.initialData.searchContent.whereType<SearchVideo>());
-    } else {
-      yield* Stream.fromIterable(page!.initialData.searchContent);
-    }
-
+      {bool onlyVideos = true, SearchFilter filter = const SearchFilter('')}) async* {
+    SearchPage? page;
     // ignore: literal_only_boolean_expressions
-    while (true) {
-      page = await page!.nextPage(_httpClient);
+    for (;;) {
       if (page == null) {
-        return;
+        page = await retry(() async => SearchPage.get(_httpClient, searchQuery, filter: filter));
+      } else {
+        page = await page.nextPage(_httpClient);
+        if (page == null) {
+          return;
+        }
       }
 
       if (onlyVideos) {
-        yield* Stream.fromIterable(
-            page.initialData.searchContent.whereType<SearchVideo>());
+        yield* Stream.fromIterable(page!.initialData.searchContent.whereType<SearchVideo>());
       } else {
-        yield* Stream.fromIterable(page.initialData.searchContent);
+        yield* Stream.fromIterable(page!.initialData.searchContent);
       }
     }
   }
@@ -69,8 +74,7 @@ class SearchClient {
   /// Queries to YouTube to get the results.
   @Deprecated('Use getVideosFromPage instead - '
       'Should be used only to get related videos')
-  Future<SearchQuery> queryFromPage(String searchQuery) =>
-      SearchQuery.search(_httpClient, searchQuery);
+  Future<SearchQuery> queryFromPage(String searchQuery) => SearchQuery.search(_httpClient, searchQuery);
 }
 
 /*